backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
google-genai>=0.8.0
python-dotenv>=1.0.0
pydantic>=2.9.0
python-docx>=0.8.11
reportlab>=4.0
python-bidi>=0.4.2
supabase>=2.0.0
pypdf>=4.0
python-multipart>=0.0.9
pdfplumber>=0.9.0
